requirements.txt
Django>=4.2.4,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.7,<3.0
drf-spectacular>=0.26.4,<0.27